requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pillow==10.2.0
pydub==0.25.1
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
requests==2.31.0
SpeechRecognition==3.10.1
typing_extensions==4.9.0
urllib3==2.2.0
